pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Abdujabbar Mirkhalikov <abdujabbar.mirkhalikov@gmail.com>"]

packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = ">=3.8"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
